Add restaurant client for managing datastore in Restaurant and Restaurants servlets.
