+++ capstone/src/main/java/com/google/step/servlets/RestaurantServlet.java
@@ -14,24 +14,17 @@
 
 package com.google.step.servlets;
 
-import com.google.appengine.api.datastore.DatastoreService;
-import com.google.appengine.api.datastore.DatastoreServiceFactory;
 import com.google.appengine.api.datastore.Entity;
 import com.google.appengine.api.datastore.GeoPt;
-import com.google.appengine.api.datastore.PreparedQuery;
-import com.google.appengine.api.datastore.Query;
-import com.google.appengine.api.datastore.Query.SortDirection;
-import com.google.gson.Gson;
-import com.google.gson.JsonObject;
-
 import com.google.appengine.api.users.UserService;
 import com.google.appengine.api.users.UserServiceFactory;
+import com.google.gson.Gson;
+import com.google.gson.JsonObject;
+import com.google.step.clients.RestaurantClient;
+import com.google.step.data.Restaurant;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Arrays;
-
-import com.google.step.data.Restaurant;
-
 import java.util.List;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -41,10 +34,6 @@ import javax.servlet.http.HttpServletResponse;
 /** Servlet responsible for sending/getting a restaurant to/from Datastore. */
 @WebServlet("/restaurant")
 public class RestaurantServlet extends HttpServlet {
-
-
-  private static DatastoreService datastore = DatastoreServiceFactory.getDatastoreService();
-
   /**
    * Will put the information gathered from the signup page and add the
    * information into their respective properties for .
@@ -105,7 +94,7 @@ public class RestaurantServlet extends HttpServlet {
     restaurantInfo.setProperty("score", 2.5);
     restaurantInfo.setProperty("status", "OKAY");
 
-    datastore.put(restaurantInfo);
+    RestaurantClient.putEntity(restaurantInfo);
 
     response.sendRedirect("/index.html");
   }
@@ -131,23 +120,13 @@ public class RestaurantServlet extends HttpServlet {
     }
    * @param request Specifies restaurant key, in the following format:
      /restaurant?restaurantKey=<int>
-   * @throws SC_NOT_FOUND if no restaurant with the requested key is found in Datastore.
    */
   @Override
   public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
     Long restaurantKeyParam = Long.parseLong(request.getParameter("restaurantKey"));
-    Query query = new Query("RestaurantInfo")
-                      .setFilter(new Query.FilterPredicate(
-                          "restaurantKey", Query.FilterOperator.EQUAL, restaurantKeyParam));
-    PreparedQuery results = datastore.prepare(query);
-    Entity resultEntity = results.asSingleEntity();
-    if (resultEntity == null) {
-      response.setStatus(HttpServletResponse.SC_NOT_FOUND);
-      return;
-    }
 
     // Restaurant object to hold all info
-    Restaurant restaurant = Restaurant.fromEntity(resultEntity);
+    Restaurant restaurant = RestaurantClient.getSingleRestaurant(restaurantKeyParam);
 
     // Format restaurant List to JSON for return
     Gson gson = new Gson();
